refactor(kb-base): 正交提取 ChunkingStrategyPanel 等 4 个内联组件和 format-utils 到 _components 目录;
将 knowledge/base/page.tsx 中的 4 个内联组件（ChunkingStrategyPanel 1491行、
CorpusSettingsPanel 364行、CorpusStatusBadge、DocumentMetadataPanel）
及 5 个工具函数（formatCorpusConfigSummary、formatFileSize、formatDateTime、
formatChunkLabel、toDocumentChunkCardViewModel）提取为独立模块，
page.tsx 从 2064 行降至 1239 行，净减 825 行。
