+++ apps/negentropy-ui/app/knowledge/base/page.tsx
@@ -6,30 +6,21 @@ import {
   useMemo,
   useRef,
   useState,
-  type Dispatch,
-  type SetStateAction,
 } from "react";
 import { usePathname, useSearchParams } from "next/navigation";
 import { toast } from "@/lib/activity-toast";
 import {
-  buildExtractorRoutesFromDraft,
-  buildCorpusConfig,
-  CorpusRecord,
-  ChunkingConfig,
-  ChunkingStrategy,
+  type CorpusRecord,
+  type ChunkingConfig,
   DocumentViewDialog,
-  DocumentChunkItem,
-  DocumentChunksMetadata,
-  ExtractorDraftRoutes,
-  ExtractorDraftTarget,
-  KnowledgeDocument,
-  KnowledgeMatch,
-  SearchMode,
-  createEmptyExtractorDraftTarget,
+  type DocumentChunkItem,
+  type DocumentChunksMetadata,
+  type KnowledgeDocument,
+  type KnowledgeMatch,
+  type SearchMode,
   fetchDocumentChunks,
   fetchDocumentChunkDetail,
   fetchDocuments,
-  normalizeExtractorDraftRoutes,
   regenerateDocumentChunkFamily,
   searchAcrossCorpora,
   updateDocumentChunk,
@@ -41,480 +32,30 @@ import {
   unarchiveDocument,
   downloadDocument,
   deleteDocument,
-  createDefaultChunkingConfig,
   normalizeChunkingConfig,
-  SeparatorsTextarea,
   PipelineStatusBadge,
-  fetchModelConfigs,
-  ModelConfigItem,
-  CorpusModelsConfig,
 } from "@/features/knowledge";
 
 import { KnowledgeNav } from "@/components/ui/KnowledgeNav";
-import { OverlayDismissLayer } from "@/components/ui/OverlayDismissLayer";
 import { outlineButtonClassName } from "@/components/ui/button-styles";
 import { AddSourceDialog } from "./_components/AddSourceDialog";
 import { CorpusFormDialog } from "./_components/CorpusFormDialog";
+import { CorpusSettingsPanel } from "./_components/CorpusSettingsPanel";
+import { CorpusStatusBadge } from "./_components/CorpusStatusBadge";
 import { DeleteCorpusDialog } from "./_components/DeleteCorpusDialog";
 import { DeleteSourceDialog } from "./_components/DeleteSourceDialog";
+import { DocumentMetadataPanel } from "./_components/DocumentMetadataPanel";
 import { RetrievedChunkCard } from "./_components/RetrievedChunkCard";
 import { RetrievedChunkDetailDialog } from "./_components/RetrievedChunkDetailDialog";
 import { EditChunkDialog } from "./_components/EditChunkDialog";
 import { ReplaceDocumentDialog } from "./_components/ReplaceDocumentDialog";
 import { buildRetrievedChunkViewModel } from "./_components/retrieved-chunk-presenter";
-import type { RetrievedChunkViewModel } from "./_components/retrieved-chunk-presenter";
+import { formatCorpusConfigSummary, toDocumentChunkCardViewModel } from "./_components/format-utils";
 
 const APP_NAME = process.env.NEXT_PUBLIC_AGUI_APP_NAME || "negentropy";
 
 type ViewMode = "overview" | "corpus";
 type CorpusTab = "documents" | "settings" | "document-chunks";
-
-function CorpusStatusBadge({ corpus }: { corpus: CorpusRecord }) {
-  const hasKnowledge = corpus.knowledge_count > 0;
-  return (
-    <span
-      className={`inline-flex items-center rounded-full px-2 py-0.5 text-[11px] font-medium ${
-        hasKnowledge
-          ? "bg-emerald-100 text-emerald-700"
-          : "bg-zinc-100 text-zinc-600"
-      }`}
-    >
-      {hasKnowledge ? "Ready" : "Empty"}
-    </span>
-  );
-}
-
-function formatCorpusConfigSummary(corpus: CorpusRecord): string {
-  const config = normalizeChunkingConfig(
-    (corpus.config ?? {}) as Record<string, unknown>,
-  );
-
-  if (config.strategy === "semantic") {
-    return `strategy: semantic · threshold: ${config.semantic_threshold.toFixed(2)} · buffer: ${config.semantic_buffer_size}`;
-  }
-
-  if (config.strategy === "hierarchical") {
-    return `strategy: hierarchical · parent: ${config.hierarchical_parent_chunk_size} · child: ${config.hierarchical_child_chunk_size}`;
-  }
-
-  return `strategy: ${config.strategy} · size: ${config.chunk_size} · overlap: ${config.overlap}`;
-}
-
-function formatFileSize(size?: number | null): string {
-  if (!size || size <= 0) return "--";
-  if (size < 1024) return `${size} B`;
-  if (size < 1024 * 1024) return `${(size / 1024).toFixed(1)} KB`;
-  return `${(size / (1024 * 1024)).toFixed(2)} MB`;
-}
-
-function formatDateTime(value?: string | null): string {
-  if (!value) return "--";
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return "--";
-  return date.toLocaleString("zh-CN");
-}
-
-function formatChunkLabel(chunk: DocumentChunkItem): string {
-  if (chunk.chunk_role === "parent") {
-    const index = chunk.parent_chunk_index ?? chunk.chunk_index;
-    return `Parent-${String(index).padStart(2, "0")}`;
-  }
-  return `Chunk-${String(chunk.chunk_index).padStart(2, "0")}`;
-}
-
-function toDocumentChunkCardViewModel(chunk: DocumentChunkItem): RetrievedChunkViewModel {
-  return {
-    id: chunk.id,
-    variant: chunk.child_chunks.length > 0 ? "hierarchical" : "standard",
-    title: formatChunkLabel(chunk),
-    characterCount: chunk.character_count,
-    preview: chunk.content,
-    fullContent: chunk.content,
-    sourceLabel: chunk.source_uri || "-",
-    sourceTitle: chunk.source_uri || "-",
-    score: 0,
-    childHitCount: chunk.child_chunks.length,
-    childChunks: chunk.child_chunks
-      .filter((child) => child.content.trim().length > 0)
-      .map((child) => ({
-        id: child.id,
-        label: `C-${String(child.child_chunk_index ?? child.chunk_index).padStart(2, "0")}`,
-        content: child.content,
-        score: 0,
-      })),
-    raw: {
-      id: chunk.id,
-      content: chunk.content,
-      source_uri: chunk.source_uri,
-      combined_score: 0,
-      metadata: chunk.metadata,
-    } as KnowledgeMatch,
-  };
-}
-
-function DocumentMetadataPanel({
-  metadata,
-}: {
-  metadata: DocumentChunksMetadata;
-}) {
-  const stats: Array<[string, string | number]> = [
-    ["Chunks specification", metadata.chunk_specification ?? "--"],
-    ["Chunks length", metadata.chunk_length ?? "--"],
-    ["Avg. paragraph length", metadata.avg_paragraph_length ?? "--"],
-    ["Paragraphs", metadata.paragraph_count ?? "--"],
-    ["Retrieval count", metadata.retrieval_count ?? 0],
-    ["Embedding time", metadata.embedding_time_ms ? `${metadata.embedding_time_ms} ms` : "--"],
-    ["Embedded spend", metadata.embedded_tokens ? `${metadata.embedded_tokens} tokens` : "--"],
-  ];
-  const docInfo: Array<[string, string | number]> = [
-    ["Original filename", metadata.original_filename ?? "--"],
-    ["Original file size", formatFileSize(metadata.file_size)],
-    ["Upload date", formatDateTime(metadata.upload_date)],
-    ["Last update date", formatDateTime(metadata.last_update_date)],
-    ["Source", metadata.source ?? "--"],
-  ];
-  const renderFieldGroup = (title: string, items: Array<[string, string | number]>) => (
-    <div>
-      <h4 className="text-[11px] font-semibold uppercase tracking-[0.24em] text-muted">{title}</h4>
-      <dl className="mt-3 space-y-2.5">
-        {items.map(([label, value]) => (
-          <div key={label} className="grid grid-cols-[minmax(0,1.2fr)_minmax(0,1.8fr)] items-start gap-x-4 gap-y-1">
-            <dt className="break-words text-sm font-semibold text-zinc-500 dark:text-zinc-400">
-              {label}
-            </dt>
-            <dd className="break-words text-sm font-medium leading-6 text-foreground">
-              {String(value)}
-            </dd>
-          </div>
-        ))}
-      </dl>
-    </div>
-  );
-
-  return (
-    <section className="h-full rounded-2xl border border-border bg-card p-5">
-      <h3 className="text-base font-semibold">Document Metadata</h3>
-      <p className="mt-2 text-sm text-muted">
-        Metadata serves as a critical filter that enhances the accuracy and relevance of information retrieval.
-      </p>
-      <div className="mt-6 space-y-6">
-        {renderFieldGroup("Document Information", docInfo)}
-        {renderFieldGroup("Technical Parameters", stats)}
-      </div>
-    </section>
-  );
-}
-
-function ChunkingStrategyPanel({
-  config,
-  onChange,
-  title,
-  description,
-}: {
-  config: ChunkingConfig;
-  onChange: Dispatch<SetStateAction<ChunkingConfig>>;
-  title: string;
-  description?: string;
-}) {
-  const strategyDescriptions: Record<ChunkingStrategy, string> = {
-    fixed: "固定长度切分，简单可预测，但可能割裂句子或段落。",
-    recursive: "按段落、句子、词递归切分，适合大多数技术文档。",
-    semantic: "基于语义相似度断点切分，完整性更高，但计算成本更高。",
-    hierarchical: "构建父子块结构，检索子块并返回父块上下文，适合长文与手册。",
-  };
-
-  const setStrategy = (strategy: ChunkingStrategy) => {
-    onChange(createDefaultChunkingConfig(strategy));
-  };
-
-  const updateConfig = (nextConfig: ChunkingConfig) => {
-    onChange(nextConfig);
-  };
-
-  return (
-    <div className="rounded-2xl border border-border bg-background p-4">
-      <div className="flex items-start justify-between gap-4">
-        <div>
-          <h3 className="text-sm font-semibold">{title}</h3>
-          {description && (
-            <p className="mt-1 text-xs text-muted">{description}</p>
-          )}
-        </div>
-        <div className="text-[11px] text-muted">大小单位: 字符近似值</div>
-      </div>
-
-      <div className="mt-3 grid gap-2 md:grid-cols-2 xl:grid-cols-4">
-        {(["fixed", "recursive", "semantic", "hierarchical"] as const).map(
-          (strategy) => (
-            <button
-              key={strategy}
-              type="button"
-              onClick={() => setStrategy(strategy)}
-              className={`rounded-xl border px-3 py-3 text-left ${
-                config.strategy === strategy
-                  ? "border-foreground bg-foreground text-background"
-                  : "border-border hover:bg-muted"
-              }`}
-            >
-              <div className="text-xs font-semibold capitalize">{strategy}</div>
-              <div
-                className={`mt-1 text-[11px] ${
-                  config.strategy === strategy ? "text-background/80" : "text-muted"
-                }`}
-              >
-                {strategyDescriptions[strategy]}
-              </div>
-            </button>
-          ),
-        )}
-      </div>
-
-      {config.strategy === "fixed" && (
-        <div className="mt-3 grid gap-3 md:grid-cols-2">
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Chunk Size</div>
-            <input
-              type="number"
-              value={String(config.chunk_size)}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  chunk_size: Number(e.target.value || 0) || 800,
-                })
-              }
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Overlap</div>
-            <input
-              type="number"
-              value={String(config.overlap)}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  overlap: Number(e.target.value || 0) || 0,
-                })
-              }
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-          <label className="inline-flex items-center gap-2 text-xs">
-            <input
-              type="checkbox"
-              checked={config.preserve_newlines}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  preserve_newlines: e.target.checked,
-                })
-              }
-            />
-            <span>保留换行</span>
-          </label>
-        </div>
-      )}
-
-      {config.strategy === "recursive" && (
-        <div className="mt-3 space-y-3">
-          <div className="grid gap-3 md:grid-cols-2 xl:grid-cols-4">
-            <label className="text-xs">
-              <div className="mb-1 text-muted">Chunk Size</div>
-              <input
-                type="number"
-                value={String(config.chunk_size)}
-                onChange={(e) =>
-                  updateConfig({
-                    ...config,
-                    chunk_size: Number(e.target.value || 0) || 800,
-                  })
-                }
-                className="w-full rounded border border-border bg-card px-2 py-2"
-              />
-            </label>
-            <label className="text-xs">
-              <div className="mb-1 text-muted">Overlap</div>
-              <input
-                type="number"
-                value={String(config.overlap)}
-                onChange={(e) =>
-                  updateConfig({
-                    ...config,
-                    overlap: Number(e.target.value || 0) || 0,
-                  })
-                }
-                className="w-full rounded border border-border bg-card px-2 py-2"
-              />
-            </label>
-            <label className="text-xs md:col-span-2">
-              <div className="mb-1 text-muted">Separators（每行一个）</div>
-              <SeparatorsTextarea
-                value={config.separators}
-                onChange={(separators) =>
-                  updateConfig({ ...config, separators })
-                }
-                rows={3}
-                placeholder={"\\n"}
-                className="w-full rounded border border-border bg-card px-2 py-2"
-              />
-            </label>
-          </div>
-          <label className="inline-flex items-center gap-2 text-xs">
-            <input
-              type="checkbox"
-              checked={config.preserve_newlines}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  preserve_newlines: e.target.checked,
-                })
-              }
-            />
-            <span>保留换行</span>
-          </label>
-        </div>
-      )}
-
-      {config.strategy === "semantic" && (
-        <div className="mt-3 grid gap-3 md:grid-cols-3">
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Similarity Threshold</div>
-            <input
-              type="number"
-              step="0.05"
-              min="0"
-              max="1"
-              value={String(config.semantic_threshold)}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  semantic_threshold: Number(e.target.value) || 0.85,
-                })
-              }
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Buffer Size</div>
-            <input
-              type="number"
-              value={String(config.semantic_buffer_size)}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  semantic_buffer_size: Number(e.target.value) || 1,
-                })
-              }
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Min Chunk Size</div>
-            <input
-              type="number"
-              value={String(config.min_chunk_size)}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  min_chunk_size: Number(e.target.value) || 50,
-                })
-              }
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Max Chunk Size</div>
-            <input
-              type="number"
-              value={String(config.max_chunk_size)}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  max_chunk_size: Number(e.target.value) || 2000,
-                })
-              }
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-        </div>
-      )}
-
-      {config.strategy === "hierarchical" && (
-        <div className="mt-3 space-y-3">
-          <div className="grid gap-3 md:grid-cols-3">
-            <label className="text-xs">
-              <div className="mb-1 text-muted">Parent Size</div>
-              <input
-                type="number"
-                value={String(config.hierarchical_parent_chunk_size)}
-                onChange={(e) =>
-                  updateConfig({
-                    ...config,
-                    hierarchical_parent_chunk_size: Number(e.target.value) || 1024,
-                  })
-                }
-                className="w-full rounded border border-border bg-card px-2 py-2"
-              />
-            </label>
-            <label className="text-xs">
-              <div className="mb-1 text-muted">Child Size</div>
-              <input
-                type="number"
-                value={String(config.hierarchical_child_chunk_size)}
-                onChange={(e) =>
-                  updateConfig({
-                    ...config,
-                    hierarchical_child_chunk_size: Number(e.target.value) || 256,
-                  })
-                }
-                className="w-full rounded border border-border bg-card px-2 py-2"
-              />
-            </label>
-            <label className="text-xs">
-              <div className="mb-1 text-muted">Child Overlap</div>
-              <input
-                type="number"
-                value={String(config.hierarchical_child_overlap)}
-                onChange={(e) =>
-                  updateConfig({
-                    ...config,
-                    hierarchical_child_overlap: Number(e.target.value) || 0,
-                  })
-                }
-                className="w-full rounded border border-border bg-card px-2 py-2"
-              />
-            </label>
-          </div>
-          <label className="text-xs">
-            <div className="mb-1 text-muted">Separators（每行一个）</div>
-            <SeparatorsTextarea
-              value={config.separators}
-              onChange={(separators) =>
-                updateConfig({ ...config, separators })
-              }
-              rows={3}
-              placeholder={"\\n"}
-              className="w-full rounded border border-border bg-card px-2 py-2"
-            />
-          </label>
-          <label className="inline-flex items-center gap-2 text-xs">
-            <input
-              type="checkbox"
-              checked={config.preserve_newlines}
-              onChange={(e) =>
-                updateConfig({
-                  ...config,
-                  preserve_newlines: e.target.checked,
-                })
-              }
-            />
-            <span>保留换行</span>
-          </label>
-        </div>
-      )}
-    </div>
-  );
-}
 
 export default function KnowledgeBasePage() {
   const pathname = usePathname();
@@ -1693,372 +1234,6 @@ export default function KnowledgeBasePage() {
         onRegenerate={() => void handleRegenerateDocumentChunkFamily()}
         pending={chunkActionPending}
       />
-    </div>
-  );
-}
-
-function CorpusSettingsPanel({
-  corpus,
-  onSave,
-}: {
-  corpus: CorpusRecord;
-  onSave: (config: Record<string, unknown>) => Promise<void>;
-}) {
-  const [formConfig, setFormConfig] = useState<ChunkingConfig>(
-    normalizeChunkingConfig((corpus.config || {}) as Record<string, unknown>),
-  );
-  const [extractorDraftRoutes, setExtractorDraftRoutes] = useState<ExtractorDraftRoutes>(
-    normalizeExtractorDraftRoutes((corpus.config || {}) as Record<string, unknown>),
-  );
-  const [servers, setServers] = useState<Array<{ id: string; name: string; display_name: string | null; is_enabled: boolean }>>([]);
-  const [toolsByServer, setToolsByServer] = useState<Record<string, Array<{ name: string; display_name: string | null; is_enabled: boolean }>>>({});
-
-  // Models settings
-  const corpusModels = ((corpus.config || {}) as Record<string, unknown>).models as CorpusModelsConfig | undefined;
-  const [llmConfigId, setLlmConfigId] = useState<string | "">(corpusModels?.llm_config_id ?? "");
-  const [embeddingConfigId, setEmbeddingConfigId] = useState<string | "">(corpusModels?.embedding_config_id ?? "");
-  const [llmModels, setLlmModels] = useState<ModelConfigItem[]>([]);
-  const [embeddingModels, setEmbeddingModels] = useState<ModelConfigItem[]>([]);
-  const [confirmDimensionDialog, setConfirmDimensionDialog] = useState<{ pending: Record<string, unknown>; newDims?: number; oldDims?: number } | null>(null);
-
-  useEffect(() => {
-    let active = true;
-
-    const loadServers = async () => {
-      const response = await fetch("/api/interface/mcp/servers");
-      if (!response.ok) {
-        throw new Error("Failed to load MCP servers");
-      }
-      const data = (await response.json()) as Array<{
-        id: string;
-        name: string;
-        display_name: string | null;
-        is_enabled: boolean;
-      }>;
-      if (!active) return;
-      const enabledServers = data.filter((item) => item.is_enabled);
-      setServers(enabledServers);
-
-      const toolEntries = await Promise.all(
-        enabledServers.map(async (server) => {
-          const toolsResponse = await fetch(`/api/interface/mcp/servers/${server.id}/tools`);
-          if (!toolsResponse.ok) {
-            return [server.id, []] as const;
-          }
-          const tools = (await toolsResponse.json()) as Array<{
-            name: string;
-            display_name: string | null;
-            is_enabled: boolean;
-          }>;
-          return [server.id, tools.filter((item) => item.is_enabled)] as const;
-        }),
-      );
-      if (!active) return;
-      setToolsByServer(Object.fromEntries(toolEntries));
-    };
-
-    void loadServers().catch((err) => {
-      toast.error(err instanceof Error ? err.message : "Failed to load MCP servers");
-    });
-
-    return () => {
-      active = false;
-    };
-  }, []);
-
-  useEffect(() => {
-    void fetchModelConfigs({ modelType: "llm", enabled: true })
-      .then(setLlmModels)
-      .catch(() => {});
-    void fetchModelConfigs({ modelType: "embedding", enabled: true })
-      .then(setEmbeddingModels)
-      .catch(() => {});
-  }, []);
-
-  const modelsConfig: CorpusModelsConfig = {
-    llm_config_id: llmConfigId || undefined,
-    embedding_config_id: embeddingConfigId || undefined,
-  };
-
-  const handleSubmit = async () => {
-    const config = buildCorpusConfig(
-      formConfig,
-      buildExtractorRoutesFromDraft(extractorDraftRoutes),
-      modelsConfig,
-    );
-
-    // 维度变更校验
-    const oldEid = corpusModels?.embedding_config_id;
-    const newEid = embeddingConfigId || undefined;
-    if (oldEid !== newEid && newEid && corpus.knowledge_count > 0) {
-      const oldItem = embeddingModels.find((m) => m.id === oldEid);
-      const newItem = embeddingModels.find((m) => m.id === newEid);
-      const oldDims = oldItem?.config?.dimensions as number | undefined;
-      const newDims = newItem?.config?.dimensions as number | undefined;
-      if (oldDims != null && newDims != null && oldDims !== newDims) {
-        setConfirmDimensionDialog({ pending: config, oldDims, newDims });
-        return;
-      }
-    }
-
-    await onSave(config);
-  };
-
-  return (
-    <div className="space-y-3">
-      <ChunkingStrategyPanel
-        config={formConfig}
-        onChange={setFormConfig}
-        title="Chunking Settings"
-        description="保存后作为该 Corpus 的默认分块配置。"
-      />
-
-      {/* Models Settings */}
-      <div className="rounded-2xl border border-border bg-background p-4">
-        <h3 className="text-sm font-semibold">Models Settings</h3>
-        <p className="mt-1 text-xs text-muted">
-          Embedding Model 影响 Embedding Indexing / Vector Search；LLM Model 影响 URL 与 PDF 文档抽取调用 MCP 时的 Plan LLM。
-        </p>
-
-        <div className="mt-3 grid grid-cols-2 gap-4">
-          <div>
-            <label className="block text-xs font-medium text-zinc-600 dark:text-zinc-400 mb-1">
-              Embedding Model
-            </label>
-            <select
-              value={embeddingConfigId}
-              onChange={(e) => setEmbeddingConfigId(e.target.value)}
-              className="w-full rounded-lg border border-zinc-200 bg-white px-3 py-2 text-sm dark:border-zinc-700 dark:bg-zinc-800 dark:text-zinc-100"
-            >
-              <option value="">(使用全局默认)</option>
-              {embeddingModels.map((m) => (
-                <option key={m.id} value={m.id}>
-                  {m.display_name} · {m.vendor}/{m.model_name}
-                </option>
-              ))}
-            </select>
-            {embeddingConfigId && (() => {
-              const sel = embeddingModels.find((m) => m.id === embeddingConfigId);
-              const dims = typeof sel?.config?.dimensions === "number" ? sel.config.dimensions : null;
-              return dims != null ? (
-                <span className="mt-1 inline-block rounded bg-blue-100 px-1.5 py-0.5 text-[10px] font-medium text-blue-700 dark:bg-blue-900/30 dark:text-blue-300">
-                  {dims} dims
-                </span>
-              ) : null;
-            })()}
-          </div>
-
-          <div>
-            <label className="block text-xs font-medium text-zinc-600 dark:text-zinc-400 mb-1">
-              LLM Model
-            </label>
-            <select
-              value={llmConfigId}
-              onChange={(e) => setLlmConfigId(e.target.value)}
-              className="w-full rounded-lg border border-zinc-200 bg-white px-3 py-2 text-sm dark:border-zinc-700 dark:bg-zinc-800 dark:text-zinc-100"
-            >
-              <option value="">(使用全局默认)</option>
-              {llmModels.map((m) => (
-                <option key={m.id} value={m.id}>
-                  {m.display_name} · {m.vendor}/{m.model_name}
-                </option>
-              ))}
-            </select>
-          </div>
-        </div>
-      </div>
-
-      {/* Dimension change confirm dialog */}
-      {confirmDimensionDialog && (
-        <OverlayDismissLayer
-          open
-          onClose={() => setConfirmDimensionDialog(null)}
-          containerClassName="p-4"
-          contentClassName="w-full max-w-sm rounded-xl border border-zinc-200 bg-white p-6 shadow-xl dark:border-zinc-700 dark:bg-zinc-900"
-          contentProps={{ role: "dialog", "aria-modal": true }}
-        >
-          <h3 className="text-base font-semibold text-red-600 dark:text-red-400 mb-2">
-            Embedding 维度变更确认
-          </h3>
-          <p className="text-sm text-zinc-700 dark:text-zinc-300">
-            切换 Embedding Model 会导致现有{" "}
-            <strong>{corpus.knowledge_count}</strong> 个 Knowledge 块的向量维度不匹配
-            （{confirmDimensionDialog.oldDims} → {confirmDimensionDialog.newDims} dims），
-            保存后系统将自动触发重建流水线，耗时取决于文档数量。
-          </p>
-          <div className="mt-4 flex justify-end gap-2">
-            <button
-              onClick={() => setConfirmDimensionDialog(null)}
-              className="px-4 py-2 rounded-lg text-xs font-medium text-zinc-600 hover:bg-zinc-100 dark:text-zinc-400 dark:hover:bg-zinc-800"
-            >
-              取消
-            </button>
-            <button
-              onClick={async () => {
-                const { pending } = confirmDimensionDialog;
-                setConfirmDimensionDialog(null);
-                await onSave(pending);
-              }}
-              className="px-4 py-2 rounded-lg text-xs font-medium bg-red-600 text-white hover:bg-red-500"
-            >
-              确认继续
-            </button>
-          </div>
-        </OverlayDismissLayer>
-      )}
-
-      <div className="rounded-2xl border border-border bg-background p-4">
-        <div className="flex items-start justify-between gap-4">
-          <div>
-            <h3 className="text-sm font-semibold">Document Extraction Settings</h3>
-            <p className="mt-1 text-xs text-muted">
-              通过 MCP Tools 为当前 Corpus 注入 URL、PDF 等源文档解释器。
-            </p>
-            <p className="mt-2 text-[11px] leading-5 text-muted">
-              可用于此处的 Tool 需提供可发现的 input/output schema，并能返回正文 Markdown 或文本；当前兼容单文档扁平协议，以及单个
-              sources 数组的 batch 协议，系统会自动把单个 URL/PDF 请求包装为对应格式。
-            </p>
-          </div>
-        </div>
-
-        {([
-          ["url", "URL 文档", "页面抓取、正文抽取、Markdown 化"],
-          ["file_pdf", "PDF 文档", "PDF 解析、Markdown 转换、图片提取"],
-        ] as const).map(([routeKey, title, description]) => {
-          const targets = extractorDraftRoutes[routeKey];
-          return (
-            <div key={routeKey} className="mt-4 rounded-xl border border-border p-3">
-              <div className="mb-3">
-                <div className="text-xs font-semibold">{title}</div>
-                <div className="text-[11px] text-muted">{description}</div>
-              </div>
-              <div className="space-y-3">
-                {[0, 1].map((index) => {
-                  const target = targets[index];
-                  const selectedServerId = target?.server_id || "";
-                  const toolOptions = selectedServerId
-                    ? (toolsByServer[selectedServerId] || [])
-                    : [];
-                  const selectedServer =
-                    servers.find((server) => server.id === selectedServerId) || null;
-                  const hasSelectedTool = toolOptions.some(
-                    (tool) => tool.name === target.tool_name,
-                  );
-
-                  const serverOptions = selectedServerId && !selectedServer
-                    ? [
-                        {
-                          id: selectedServerId,
-                          name: selectedServerId,
-                          display_name: "已配置 MCP（当前不可用）",
-                          is_enabled: false,
-                        },
-                        ...servers,
-                      ]
-                    : servers;
-                  const visibleToolOptions =
-                    target.tool_name && !hasSelectedTool
-                      ? [
-                          {
-                            name: target.tool_name,
-                            display_name: "已配置 Tool（当前不可用）",
-                            is_enabled: false,
-                          },
-                          ...toolOptions,
-                        ]
-                      : toolOptions;
-
-                  const setTarget = (patch: Partial<ExtractorDraftTarget>) => {
-                    setExtractorDraftRoutes((prev) => {
-                      const nextRoute = [...prev[routeKey]] as typeof prev[typeof routeKey];
-                      nextRoute[index] = {
-                        ...nextRoute[index],
-                        ...patch,
-                        priority: index,
-                        enabled: true,
-                      };
-                      return {
-                        ...prev,
-                        [routeKey]: nextRoute,
-                      };
-                    });
-                  };
-
-                  const clearTarget = () => {
-                    setExtractorDraftRoutes((prev) => {
-                      const nextRoute = [...prev[routeKey]] as typeof prev[typeof routeKey];
-                      nextRoute[index] = createEmptyExtractorDraftTarget(index);
-                      return {
-                        ...prev,
-                        [routeKey]: nextRoute,
-                      };
-                    });
-                  };
-
-                  return (
-                    <div key={`${routeKey}-${index}`} className="grid gap-3 rounded-lg border border-border bg-card p-3 md:grid-cols-[120px_1fr_1fr_auto]">
-                      <div className="text-xs font-semibold text-zinc-600 dark:text-zinc-300">
-                        {index === 0 ? "主用" : "备用"}
-                      </div>
-                      <label className="text-xs">
-                        <div className="mb-1 text-muted">MCP Server</div>
-                        <select
-                          value={selectedServerId}
-                          onChange={(e) =>
-                            setTarget({ server_id: e.target.value, tool_name: "" })
-                          }
-                          className="w-full rounded border border-border bg-background px-2 py-2"
-                        >
-                          <option value="">未配置</option>
-                          {serverOptions.map((server) => (
-                            <option key={server.id} value={server.id}>
-                              {server.display_name || server.name}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                      <label className="text-xs">
-                        <div className="mb-1 text-muted">Tool</div>
-                        <select
-                          value={target?.tool_name || ""}
-                          onChange={(e) => setTarget({ tool_name: e.target.value })}
-                          disabled={!selectedServerId}
-                          className="w-full rounded border border-border bg-background px-2 py-2 disabled:opacity-50"
-                        >
-                          <option value="">未配置</option>
-                          {visibleToolOptions.map((tool) => (
-                            <option key={tool.name} value={tool.name}>
-                              {tool.display_name || tool.name}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                      <div className="flex items-end">
-                        <button
-                          type="button"
-                          onClick={clearTarget}
-                          className={outlineButtonClassName("neutral", "rounded px-3 py-2 text-[11px]")}
-                        >
-                          清空
-                        </button>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          );
-        })}
-      </div>
-
-      <div className="flex justify-end">
-        <button
-          onClick={handleSubmit}
-          className="rounded bg-zinc-900 px-3 py-2 text-xs font-semibold text-white"
-        >
-          Save Settings
-        </button>
-      </div>
     </div>
   );
 }
